backend/requirements.txt
fastapi
uvicorn
python-socketio
pydantic
pydantic-settings
deepgram-sdk==2.12.0
groq
